dafrjulia: R Interface to the Julia 'DataAxesFormats' Package

An R interface to 'DataAxesFormats.jl' https://github.com/tanaylab/DataAxesFormats.jl, a 'Julia' https://julialang.org/ package providing a uniform interface for accessing multi-dimensional data arranged along arbitrary axes. This package generalizes and extends 'AnnData'-like functionality, supporting both 1D and 2D data, multiple data modalities, and efficient memory management. 'dafrjulia' enables R users to create, read, modify, and analyze complex biological datasets such as single-cell data with multiple modalities (RNA-seq, ATAC-seq, etc.). The package implements efficient memory management strategies and provides a seamless interface between R and 'Julia' for handling large-scale biological data.

See also

Author

Maintainer: Aviezer Lifshitz aviezer.lifshitz@weizmann.ac.il (ORCID)

Authors:

Oren Ben-Kiki oren@ben-kiki.org

Other contributors:

Weizmann Institute of Science [copyright holder]